+++ juriscraper/opinions/united_states/federal_appellate/ca11_p.py
@@ -55,12 +55,12 @@ class Site(OpinionSite):
             e for e in self.html.xpath("//tr[./td[1]/a//text()]/td[4]//text()")
         ]
 
-    def _download_backwards(self, n):
+    async def _download_backwards(self, n):
         self.url = "http://media.ca11.uscourts.gov/opinions/pub/logname.php?begin={}&num={}&numBegin=1".format(
             n, n / 20 - 1
         )
 
-        self.html = self._download()
+        self.html = await self._download()
         if self.html is not None:
             # Setting status is important because it prevents the download
             # function from being run a second time by the parse method.