+++ juriscraper/opinions/united_states/state/miss.py
@@ -29,8 +29,8 @@ class Site(OpinionSiteLinear):
     them.
     """
 
-    def _download(self, request_dict={}):
-        dates_page = super()._download(request_dict)
+    async def _download(self, request_dict={}):
+        dates_page = await super()._download(request_dict)
         self.parse_date_pages(dates_page)
 
     """Keep track of the most recent N date pages.